// src/client/models/MaterialOperationTypeEnum.ts
/* generated using openapi-typescript-codegen -- do no edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */

/**
 * Material operation type.
 *
 * - FEED: loading material
 * - UNFEED: unloading material
 */
export enum MaterialOperationTypeEnum {
    FEED = 'FEED',
    UNFEED = 'UNFEED',
}