// src/client/models/ProduceTypeEnum.ts
/* generated using openapi-typescript-codegen -- do no edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */

/**
 * - NORMAL_PRODUCE_MODE: 正式生產模式
 * - TESTING_PRODUCE_MODE: 量試生產模式
 */
export enum ProduceTypeEnum {
    NORMAL_PRODUCE_MODE = 'NORMAL_PRODUCE_MODE',
    TESTING_PRODUCE_MODE = 'TESTING_PRODUCE_MODE',
}